Manage dangling soft links
If a file in our snap is already available in an extension snap
(like, for example, gnome-46-2040), but there are soft links to
it in our snap that aren't available in the extension snap, the
soft link will become dangling.

This patch fixes this by ensuring that deleting a file with
soft links will before transform all the soft links into
hard links, and only then removing the original file. This
ensures that soft links will remain accessible.